+++ kpops/components/base_components/pipeline_component.py
@@ -15,10 +15,12 @@ from kpops.components.base_components.models.from_section import (
 )
 from kpops.components.base_components.models.resource import Resource
 from kpops.components.base_components.models.to_section import (
+    ToSection,
+)
+from kpops.components.base_components.models.topic import (
     KafkaTopic,
     OutputTopicTypes,
     TopicConfig,
-    ToSection,
 )
 from kpops.utils.docstring import describe_attr
