+++ tests/component_handlers/topic/test_topic_handler.py
@@ -19,7 +19,7 @@ from kpops.component_handlers.topic.model import (
     TopicResponse,
     TopicSpec,
 )
-from kpops.components.base_components.models.to_section import (
+from kpops.components.base_components.models.topic import (
     KafkaTopic,
     OutputTopicTypes,
     TopicConfig,
