+++ tests/components/test_kafka_source_connector.py
@@ -23,10 +23,9 @@ from kpops.components.base_components.models.from_section import (
     TopicName,
 )
 from kpops.components.base_components.models.to_section import (
-    OutputTopicTypes,
-    TopicConfig,
     ToSection,
 )
+from kpops.components.base_components.models.topic import OutputTopicTypes, TopicConfig
 from kpops.config import KpopsConfig
 from kpops.utils.environment import ENV
 from tests.components.test_kafka_connector import (
